pyproject.toml
[project]
name = "harmonic-compression"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "glfw>=2.8.0",
    "imgui>=2.0.0",
    "numba>=0.61.0",
    "open3d>=0.19.0",
    "opencv-python>=4.11.0.86",
    "pye57>=0.4.16",
    "pyopengl>=3.1.9",
    "pyopengl-accelerate>=3.1.9",
    "quaternionic>=1.0.15",
    "scipy>=1.15.2",
    "spherical>=1.0.15",
]
